// src/types/component-types-sb.d.ts
// This file was generated by the storyblok CLI.
// DO NOT MODIFY THIS FILE BY HAND.
import type { ISbStoryData } from "storyblok";
export interface RichtextStoryblok {
  type: string;
  content?: RichtextStoryblok[];
  marks?: RichtextStoryblok[];
  attrs?: any;
  text?: string;
  [k: string]: any;
}

export interface AboutSectionStoryblok {
  title: string;
  description: RichtextStoryblok;
  skills?: AboutSectionSkillItemStoryblok[];
  usp_items?: AboutSectionUspItemStoryblok[];
  component: "AboutSection";
  _uid: string;
  [k: string]: any;
}

export interface AboutSectionSkillItemStoryblok {
  label: string;
  component: "AboutSectionSkillItem";
  _uid: string;
  [k: string]: any;
}

export interface AboutSectionUspItemStoryblok {
  icon: string;
  title: string;
  subtitle: string;
  component: "AboutSectionUspItem";
  _uid: string;
  [k: string]: any;
}

export interface ContactSectionStoryblok {
  title: string;
  subtitle: string;
  items?: ContactSectionItemStoryblok[];
  component: "ContactSection";
  _uid: string;
  [k: string]: any;
}

export interface ContactSectionItemStoryblok {
  icon: string;
  title: string;
  subtitle: string;
  url: string;
  component: "ContactSectionItem";
  _uid: string;
  [k: string]: any;
}

export interface AssetStoryblok {
  alt: string | null;
  copyright?: string | null;
  fieldtype: "asset";
  id: number;
  filename: string | null;
  name: string;
  title: string | null;
  focus: string | null;
  meta_data?: {
    [k: string]: any;
  };
  source?: string | null;
  is_external_url?: boolean;
  is_private?: boolean;
  src?: string;
  updated_at?: string;
  width?: number | null;
  height?: number | null;
  aspect_ratio?: number | null;
  public_id?: string | null;
  content_type?: string;
  [k: string]: any;
}

export interface HeroHeaderStoryblok {
  title: string;
  subtitle: string;
  button_label: string;
  button_url: string;
  trusted_by: string;
  image?: AssetStoryblok;
  component: "HeroHeader";
  _uid: string;
  [k: string]: any;
}

export interface HeroSectionStoryblok {
  video_cover?: AssetStoryblok;
  video_url?: string;
  title: string;
  description: string;
  expertises?: HeroSectionExpertiseItemStoryblok[];
  component: "HeroSection";
  _uid: string;
  [k: string]: any;
}

export interface HeroSectionExpertiseItemStoryblok {
  icon: string;
  label: string;
  component: "HeroSectionExpertiseItem";
  _uid: string;
  [k: string]: any;
}

export interface LargeVideoStoryblok {
  video?: AssetStoryblok;
  component: "LargeVideo";
  _uid: string;
  [k: string]: any;
}

export interface PageStoryblok {
  body?: (
    | AboutSectionStoryblok
    | ContactSectionStoryblok
    | HeroHeaderStoryblok
    | HeroSectionStoryblok
    | LargeVideoStoryblok
    | ProjectsGridStoryblok
    | ServicesGridStoryblok
  )[];
  component: "page";
  _uid: string;
  [k: string]: any;
}

export interface ProjectsGridStoryblok {
  projects?: ProjectsGridItemStoryblok[];
  component: "ProjectsGrid";
  _uid: string;
  [k: string]: any;
}

export interface ProjectsGridItemStoryblok {
  title?: string;
  cover?: AssetStoryblok;
  year?: string;
  duration_in_seconds?: string;
  tags?: ("" | "R&D" | "Narrative" | "AI" | "Edit" | "Color" | "Docu" | "VFX" | "Promo" | "Production")[];
  client?: string;
  role?: ("" | "Creative Director" | "Director" | "Writer" | "Editor" | "Head of Post" | "AI Artist")[];
  created_at?: string;
  video_url?: string;
  video_id: string;
  studio?: string;
  platform?: string;
  component: "ProjectsGridItem";
  _uid: string;
  [k: string]: any;
}

export interface ServicesGridStoryblok {
  title: string;
  subtitle: string;
  services?: ServicesGridItemStoryblok[];
  component: "ServicesGrid";
  _uid: string;
  [k: string]: any;
}

export interface ServicesGridItemStoryblok {
  icon: string;
  title: string;
  description: string;
  component: "ServicesGridItem";
  _uid: string;
  [k: string]: any;
}